requirements.txt
playwright==1.40.0
pillow==10.1.0
pydantic==2.5.0
loguru==0.7.2
pyyaml==6.0.1
pandas==2.1.4
openpyxl==3.1.2
